requirements.txt
# Monday BI Agent - Dependencies
# Python 3.11+

# --- UI ---
streamlit>=1.32.0

# --- Data Processing ---
pandas>=2.1.0
numpy>=1.24.0

# --- Analytics Engine ---
duckdb>=0.10.0

# --- AI / LLM ---
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0

# --- Visualization ---
plotly>=5.18.0

# --- Monday.com API ---
requests>=2.31.0

# --- Utilities ---
python-dotenv>=1.0.0
